feat: Refactor chat functionality by consolidating chat panels into a single ChatPanel; enhance message handling and UI for better user experience; update DatabaseAccessor for improved chat room management.
