+++ src/main/java/com/example/OnlineMedicalAppointment/database/DatabaseAccessor.java
@@ -488,6 +488,7 @@ public abstract class DatabaseAccessor {
                     roomIds.add(rs.getString("roomID"));
                 }
             }
+            roomIds.removeFirst();
         } catch (SQLException e) {
             System.out.println("SQL Exception thrown while getting chat room IDs: " + e.getMessage());
             return new ArrayList<>(); // Return empty list on error
