+++ src/main/java/com/example/OnlineMedicalAppointment/model/ChatRoom.java
@@ -10,21 +10,40 @@ import com.example.OnlineMedicalAppointment.database.DatabaseAccessor;
  */
 public class ChatRoom {
     private String chatRoomName;
-    private int chatRoomID;
+    private String chatRoomID;
     private User chattingWith;
-    private List<Message> messages;
     
     public ChatRoom(String chatRoomID, User currentUser) {
-
-        String [] users = chatRoomID.split("-");
-            if(currentUser.getUserID() == Integer.parseInt(users[0])) {
-                User messagingWith = DatabaseAccessor.getUserByID(Integer.parseInt(users[1]));
-                chatRoomName = messagingWith.getFName() + " " + messagingWith.getLName();
+        this.chatRoomID = chatRoomID;
+        if (chatRoomID == null || chatRoomID.trim().isEmpty() || !chatRoomID.contains("-")) {
+            // Defensive: invalid chatRoomID format, avoid NumberFormatException
+            chattingWith = null;
+            chatRoomName = "Unknown User";
+            return;
+        }
+        String[] users = chatRoomID.split("-");
+        if (users.length != 2) {
+            chattingWith = null;
+            chatRoomName = "Unknown User";
+            return;
+        }
+        try {
+            int userId0 = Integer.parseInt(users[0]);
+            int userId1 = Integer.parseInt(users[1]);
+            if (currentUser.getUserID() == userId0) {
+                chattingWith = DatabaseAccessor.getUserByID(userId1);
             } else {
-                User messagingWith = DatabaseAccessor.getUserByID(Integer.parseInt(users[0]));
-                chatRoomName = messagingWith.getFName() + " " + messagingWith.getLName(); 
+                chattingWith = DatabaseAccessor.getUserByID(userId0);
             }
-            messages =  DatabaseAccessor.getMessages(chatRoomID);
+            if (chattingWith != null) {
+                chatRoomName = chattingWith.getFName() + " " + chattingWith.getLName();
+            } else {
+                chatRoomName = "Unknown User";
+            }
+        } catch (NumberFormatException e) {
+            chattingWith = null;
+            chatRoomName = "Unknown User";
+        }
     } 
     public String getChatRoomName() {
         return chatRoomName;
@@ -34,11 +53,11 @@ public class ChatRoom {
         this.chatRoomName = chatRoomName;
     }
 
-    public int getChatRoomID() {
+    public String getChatRoomID() {
         return chatRoomID;
     }
 
-    public void setChatRoomID(int chatRoomID) {
+    public void setChatRoomID(String chatRoomID) {
         this.chatRoomID = chatRoomID;
     }
 
@@ -49,9 +68,8 @@ public class ChatRoom {
         this.chattingWith = chattingWith;
     }
     public List<Message> getMessages() {
+        List<Message> messages = DatabaseAccessor.getMessages(chatRoomID);
+        System.out.println("ChatRoom.getMessages: " + messages);
         return messages;
-    }
-    public void setMessages(List<Message> messages) {
-        this.messages = messages;
     }
 }
